requirements.txt
Django==1.7.5
Pillow==2.7.0
PyQRCode==1.0
argparse==1.2.1
django-bootstrap3==5.1.1
django-simple-captcha==0.4.4
pypng==0.0.17
pytz==2014.10
requests==2.5.3
six==1.9.0
wsgiref==0.1.2
